+++ pkg/utils/database/postgres.go
@@ -23,7 +23,7 @@ import (
 type Postgres struct {
 	Backend    string
 	Host       string
-	Port       int32
+	Port       uint16
 	Database   string
 	User       string
 	Password   string
